fix redirects
